pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.3.3",
    "pdf2image>=1.17.0",
    "pillow>=12.0.0",
    "poppler-utils>=0.1.0",
    "pymupdf>=1.26.6",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-dateutil>=2.9.0.post0",
    "streamlit>=1.51.0",
]
